+++ app/layout.tsx
@@ -4,9 +4,9 @@ import type { Metadata } from "next";
 import { Black_Han_Sans, Noto_Sans_KR } from "next/font/google";
 import localFont from "next/font/local";
 
+import { Header } from "../components/Header";
+import { Providers } from "../components/Providers";
 import { classNames } from "../utils/classNames";
-import { Header } from "./_feature/components/Header";
-import { Providers } from "./_feature/Providers";
 
 const notoSansKr = Noto_Sans_KR({
   subsets: ["latin"],
