+++ app/page.tsx
@@ -5,9 +5,9 @@ import { PropsWithChildren } from "react";
 import { getNotes } from "@/notion/getNotes";
 import { getRecords } from "@/notion/getRecords";
 
-import { Divider } from "../components/Divider";
-import { Title } from "./_feature/components/Title";
-import { NoteListCard } from "./note/_feature/components/NoteListCard";
+import { Divider } from "../components/common/Divider";
+import { NoteListCard } from "../components/note/NoteListCard";
+import { Title } from "../components/Title";
 
 const SectionTitle = ({
   href,
