+++ components/note/CategoryNav.tsx
