+++ components/note/NoteListCard.tsx
