+++ components/record/RecordCard.tsx
